New snapping mechanism: icons get 'replace' sites, which highlight (in red) the resulting icons to be replaced on drop, when snapped.  The replacement target can also be modified by holding the shift key when snapped, and dragging 1) left and up to expand the selection structurally (up the tree), or 2) right or down to lexically extend the selection from the icon.  Snapping directly on a cursor site inserts, rather than replacing, but icons with series sites also get series insert sites, which always add to the series (to distinguish from appending or prepending to the existing expressions).
Also, further formalizes the concept of a "partId" to identify the physically independent (as well as lexically disjoint) parts of an icon.  This was already used for keeping track of what part of the icon the user picked up, but is now also used to figure out what part of an icon is within a selected lexical range.  It also allows for the replacement of a lot of icon-specific code, in the form of offsetOfPart and touchesPosition methods, with a single method for correlating drawList elements with partIds.

This commit still behaves badly for some cases involving empty sites, as it has become clear that selections should be able to represent them.  However, that will have to come in a future commit, as this one is already very large.

Removed the capability to snap backwards on to an open output site, which only worked under very weird and limited conditions, and more generally, restricted snapping to the top-level dragged icon from where the user pressed the mouse to start the drag.  This was changed from snapping any mating icons, because the main purpose for dragging multiple sequences is for relocation, and having random unexpected icons snap together makes that worse.

Removed the capability to snap comments to any site in a statement.  This, I think was inadvertent, and too much snapping where you wouldn't expect to attach something feels confusing.

Bugs Fixed:
 * Geometric traversal out of an entry icon to one of its sequence sites could put the cursor on a deleted icon when the act of focusing out would cause its removal.
 * Infix icons were assumed never to be nested, but unfortunately can end up that way.  Give them leftAssoc and rightAssoc methods to avoid a crash if they do.
 * insertIconsFromSequences was making some incorrect calls about not being able to process insertions because isRightmostSite was calculated wrong for cases where the site was seqOut.